added recycler view with pagination and animation, add sorting

+++ app/src/main/java/com/go0ose/cryptocurrencyapp/data/retrofit/CryptoResponse.kt
@@ -10,7 +10,7 @@ data class CryptoResponse(
     @SerializedName("image")
     val image: String,
     @SerializedName("market_cap")
-    val marketCap: Long,
+    val marketCap: Double,
     @SerializedName("name")
     val name: String,
     @SerializedName("symbol")
