+++ app/src/main/java/com/go0ose/cryptocurrencyapp/presentation/model/Coin.kt
@@ -6,5 +6,5 @@ data class Coin(
     val symbol: String,
     val currentPrice: Double,
     val image: String,
-    val marketCap: Long
+    val marketCap: Double
 )
